+++ src/audio/Music.ts
@@ -131,6 +131,10 @@ export class Music extends AbstractMedia {
         return this;
     }
 
+    protected override _getAudioSetup(): { audioContext: AudioContext; gainNode: GainNode } | null {
+        return this._audioSetup;
+    }
+
     public override destroy(): void {
         super.destroy();